[80] extract socket io event handlers and error handlers.
